fix: Правки после код-ревью
