+++ src/components/private-route/private-route.tsx
@@ -1,12 +1,15 @@
 import { AuthorizationStatus, AppRoute } from '../../const';
 import { Navigate } from 'react-router-dom';
+import { useAppSelector } from '../../hooks';
+import { selectAuthorizationStatus } from '../../store/selectors';
 
 type PrivateRouteProps = {
-  authorizationStatus: AuthorizationStatus;
   children: JSX.Element;
 }
 
-function PrivateRoute({authorizationStatus, children}:PrivateRouteProps):JSX.Element {
+function PrivateRoute({children}:PrivateRouteProps):JSX.Element {
+  const authorizationStatus = useAppSelector(selectAuthorizationStatus);
+
   return (
     authorizationStatus === AuthorizationStatus.Auth
       ? children
