+++ src/index.tsx
@@ -4,7 +4,7 @@ import App from './components/app/app';
 import comments from './mocks/comments';
 import { Provider } from 'react-redux';
 import { store } from './store';
-import { checkAuthStatus, fetchOffersAction } from './store/api-actions';
+import { checkAuthAction, fetchOffersAction } from './store/api-actions';
 import { initLocalStorage } from './services/local-storage';
 import ErrorMessage from './components/error-message/error-message';
 
@@ -13,7 +13,7 @@ const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 
-checkAuthStatus();
+store.dispatch(checkAuthAction());
 store.dispatch(fetchOffersAction());
 //Забираем данные из localStorage и помещаем в хранилище, если они есть
 initLocalStorage();
