+++ src/store/reducer.ts
@@ -1,7 +1,8 @@
 import { createReducer } from '@reduxjs/toolkit';
 import { AuthorizationStatus, CITIES, SortOption } from '../const';
 import { OffersData } from '../types/offers';
-import { changeActiveOfferId, changeCity, changeSortOffers, loadOffers, requireAuthorization, setError, setOffersDataLoadingStatus } from './action';
+import { changeActiveOfferId, changeCity, changeSortOffers, requireAuthorization, setError } from './action';
+import { checkAuthAction, fetchOffersAction } from './api-actions';
 
 const initialState = {
   currentCity: CITIES[0],
@@ -18,8 +19,18 @@ const reducer = createReducer(initialState, (builder)=> {
     .addCase(changeCity,(state, action) => {
       state.currentCity = action.payload;
     })
-    .addCase(loadOffers,(state, action) => {
+    .addCase(fetchOffersAction.pending, (state) => {
+      state.isOffersDataLoading = true;
+    })
+    .addCase(fetchOffersAction.fulfilled,(state, action) => {
       state.offers = action.payload;
+      state.isOffersDataLoading = false;
+    })
+    .addCase(checkAuthAction.rejected, (state) => {
+      state.authorizationStatus = AuthorizationStatus.NoAuth;
+    })
+    .addCase(checkAuthAction.fulfilled, (state) => {
+      state.authorizationStatus = AuthorizationStatus.Auth;
     })
     .addCase(changeSortOffers, (state, action) => {
       state.currentSortOffers = action.payload;
@@ -30,13 +41,9 @@ const reducer = createReducer(initialState, (builder)=> {
     .addCase(requireAuthorization, (state, action) => {
       state.authorizationStatus = action.payload;
     })
-    .addCase(setOffersDataLoadingStatus, (state, action) => {
-      state.isOffersDataLoading = action.payload;
-    })
     .addCase(setError, (state, action) => {
       state.error = action.payload;
     });
-
 });
 
 export {reducer};
